Implement closing a stream
This means we can restart a stream without having to kill the server and
start it again, which is convenient for development purposes.
Furthermore, the goal is for multiple clients to be able to connect to
the same server and receive a stream, which could be useful if for
example each client connects to different vGPU VMs (etc).