Fix mouse lag
We used to have a mouse lag of about 10 seconds. This is because the
channel was unbounded, so the UI could push an inordinate amount of
mouse events to be handled by the blocking client. Unfortunately, this
means that there would be a huge backlog of mouse messages, causing the
mouse lag to increase too much. A smaller ringbuffer size would fix
this by dropping mouse events, but this was not working properly due to
a bounded client response channel, which fills up immediately if both
UI -> client and client -> UI channel pairs are bounded to 1 element.
The client panicked on a fail to send the UI thread its messages, which
thereby caused the entire client handler to crash and no longer accept
events. We want to ultimately have the UI thread handle all messages,
but in the future perhaps we do not send thousands of `Ok` messages because
then the UI thread will lag.

I do not think responses are handled by the ChannelUI component yet,
which creates a memory leak, but we will handle those later. We do not
care about the latency of response messages as much as the
responsiveness of client -> server, so this solution (of an unbounded
response channel) is acceptable at the moment.